requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-filter==24.3
django-cors-headers==4.6.0
psycopg[binary]==3.2.4
python-decouple==3.8
django-import-export==4.3.3
celery==5.4.0
redis==5.2.1
gunicorn==23.0.0
whitenoise==6.8.2
Pillow==11.1.0
